pyproject.toml
[project]
name = "sf-data-pipelines"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.37.1",
    "connectorx>=0.4.3",
    "cvxpy>=1.6.4",
    "exchange-calendars>=4.9",
    "fsspec>=2025.7.0",
    "ipykernel>=6.29.5",
    "numpy>=2.2.3",
    "pandas-market-calendars>=5.1.1",
    "polars>=1.24.0",
    "pyarrow>=19.0.1",
    "python-dotenv>=1.0.1",
    "ray>=2.44.1",
    "rich>=14.1.0",
    "s3fs>=2025.7.0",
    "tqdm>=4.67.1",
    "wrds>=3.3.0",
]
